Add files via upload
The thermal conductivities of the microstructure pack which have been generated by packing software "Rocpack", and homogenized using an FFT algorithm to find out their effective thermal conductivity tensors. The packs were almost isotropic, so the effective thermal conductivity can be a scalar which is the average of the diagonal elements of the conductivity tensors. The files can be opened with Notepad or similar text editors.